* Fixed : Comments were editable / approvable by any user who could edit the associated post, regardless of moderate_comments capability
* Fixed : Terms could not be managed based on term-assigned Manager role.

* Fixed : Image / attachment access fails due to corrupted uploads/.htaccess file, under unusual editing conditions

* Change: Further safeguard against corruption of main .htaccess

* Perf  : Avoid redundant regeneration of uploads/.htaccess when attaching a file in Post Edit Form

* Perf  : Now fully utilizing WP_Comment_Query hooks introduced in WP 3.1

* Perf  : Now fully utilizing WP_Term_Query hooks introduced in WP 3.1. To revert to previous filtering, define('SCOPER_LEGACY_TERMS_FILTER');

* Perf  : Eliminated several redundant DB queries

* Perf  : Removed some extra wp-admin query filtering which is no longer necessary with WP > 3.1.  To revert to previous filtering, define('SCOPER_LEGACY_HW_FILTERS');

git-svn-id: https://plugins.svn.wordpress.org/role-scoper/trunk@419383 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ mu-init_rs.php
@@ -132,8 +132,8 @@ function scoper_establish_group_scope() {
 
 
 function scoper_review_file_htaccess() {
-	$min_date = scoper_get_site_option( 'file_htaccess_min_date' );
-	$last_regen = scoper_get_option( 'file_htaccess_date' );
+	$min_date = (int) scoper_get_site_option( 'file_htaccess_min_date' );
+	$last_regen = (int) scoper_get_option( 'file_htaccess_date' );
 
 	if ( ! $last_regen || ( $min_date > $last_regen ) )
 		scoper_flush_file_rules();
